requirements.txt
# Core ML & DL
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Explainability - using custom implementation in explainability/gradcam.py

# MLOps
mlflow>=2.8.0
dvc>=3.30.0
dvc[s3]>=3.30.0
dagshub>=0.3.0

# Model Optimization
onnx>=1.14.0
onnxruntime>=1.15.0
# onnx-simplifier>=0.4.0  # Optional - has build issues, install separately if needed

# API (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
python-dotenv>=1.0.0
tensorboard>=2.13.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
